requirements.txt
python-dotenv
pyTelegramBotApi
aiohttp